Extract common test code to IngestJobQueueConsumerTestBase
